Code Cleaning

+++ Progemon/src/logic/terrain/FightTerrain.java
@@ -14,7 +14,7 @@ import graphic.IRenderable;
 public class FightTerrain implements IRenderable {
 	
 	public final static int IMG_SIZE_X = 40, IMG_SIZE_Y = 40;
-	private int x, y;
+	private short x, y;
 	private boolean isShadowed, isCursor, isPathSign, isHighlight;
 	private TerrainType type;
 	private BufferedImage terrainImage = null;
